requirements.txt
python-telegram-bot==20.3
motor
python-dotenv